+++ CHANGELOG.md
@@ -22,6 +22,8 @@ after disconnection (_un-registration_).
 
 - Added `rpc_pending_num` metric. ([@palkan][])
 
+- Add statsd support for metrics. ([@jeremycw][])
+
 ## 1.0.2 (2020-09-08)
 
 - Add channel states to `disconnect` requests. ([@palkan][])
@@ -100,3 +102,4 @@ See [Changelog](https://github.com/anycable/anycable-go/blob/0-6-stable/CHANGELO
 [@bibendi]: https://github.com/bibendi
 [@rolandg]: https://github.com/rolandg
 [@gr8bit]: https://github.com/gr8bit
+[@jeremycw]: https://github.com/jeremycw